4.23 done

+++ osa4/blog_list/controllers/blogs.js
@@ -1,60 +1,61 @@
-const blogsRouter = require('express').Router()
-const Blog = require('../models/blog')
-const User = require('../models/user')
-const { nonExistingId } = require('../tests/test_helper')
+const router = require('express').Router()
 const jwt = require('jsonwebtoken')
 
-blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog
+const Blog = require('../models/blog')
+const User = require('../models/user')
+
+router.get('/', async (request, response) => {
+  const notes = await Blog
+    .find({})
     .find({}).populate('user', { username: 1, name: 1 })
 
-  response.json(blogs)
-});
-
-blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  response.json(notes)
 })
 
-blogsRouter.post('/', async (request, response) => {
-  const body = request.body
-  const token = getTokenFrom(request)
-  const decodedToken = jwt.verify(token, process.env.SECRET)
-  if (!decodedToken.id) {
+router.post('/', async (request, response) => {
+  if (!request.user) {
     return response.status(401).json({ error: 'token missing or invalid' })
   }
-  const user = await User.findById(body.userId)
 
-  const blog = new Blog({
-    title: body.title,
-    author: body.author,
-    url: body.url,
-    likes: body.likes,
-    user: user._id
-  })
+  const user = request.user
+  const blog = new Blog({ ...request.body, user: user.id })
 
   const savedBlog = await blog.save()
+
   user.blogs = user.blogs.concat(savedBlog._id)
   await user.save()
 
-  response.json(savedBlog)
+  response.status(201).json(savedBlog)
 })
 
-blogsRouter.put('/:id', async (request, response) => {
-  console.log(request.body)
-  Blog.findByIdAndUpdate(request.params.id, request.body, {new: true})
-    .then(updatedBlog => {
-      response.json(updatedBlog)
-    })
-    .catch(error => nonExistingId(error))
-})
-
-const getTokenFrom = request => {
-  const authorization = request.get('authorization')
-  if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
-    return authorization.substring(7)
+router.delete('/:id', async (request, response) => {
+  const blogToDelete = await Blog.findById(request.params.id)
+  if (!blogToDelete ) {
+    return response.status(204).end()
   }
-  return null
-}
 
-module.exports = blogsRouter
+  if ( blogToDelete.user && blogToDelete.user.toString() !== request.user.id ) {
+    return response.status(401).json({
+      error: 'only the creator can delete a blog'
+    })
+  }
+
+  await Blog.findByIdAndRemove(request.params.id)
+
+  response.status(204).end()
+})
+
+router.put('/:id', async (request, response) => {
+  const blog = request.body
+
+  const updatedBlog = await Blog
+    .findByIdAndUpdate(
+      request.params.id, 
+      blog, 
+      { new: true, runValidators: true, context: 'query' }
+    )
+      
+  response.json(updatedBlog)
+})
+
+module.exports = router
